+++ ricaun.RevitTest.Application/Revit/Utils/TestExecuteUtils.cs
@@ -18,7 +18,7 @@ namespace ricaun.RevitTest.Application.Revit
         {
             Log.WriteLine($"TestExecuteUtils: {filePath}");
             var zipFile = await revitTask.Run(() => CopyFilesUsingZipFolder(filePath));
-            
+
             if (zipFile is null)
             {
                 Log.WriteLine($"TestExecuteUtils: Copy Zip Fail");
@@ -27,15 +27,14 @@ namespace ricaun.RevitTest.Application.Revit
 
             string zipDestination = null;
             var extractToTempSucess = await revitTask.Run(() => ZipExtension.ExtractToTempFolder(zipFile, out zipDestination));
-            
+
             if (extractToTempSucess == false)
             {
                 Log.WriteLine($"TestExecuteUtils: Extract Zip Fail");
                 throw new Exception("Extract Zip Fail");
             }
 
-            //TestAssemblyModel tests = await revitTask.Run(() => TestDirectory(zipDestination, parameters));
-            TestAssemblyModel tests = await TestDirectoryAsync(revitTask, zipDestination, Path.GetFileName(filePath), parameters);
+            TestAssemblyModel tests = await Task.Run(() => TestDirectoryAsync(revitTask, zipDestination, Path.GetFileName(filePath), parameters));
 
             await revitTask.Run(() => CopyFilesBackUsingZip(filePath, zipDestination));
 
@@ -85,9 +84,9 @@ namespace ricaun.RevitTest.Application.Revit
         {
             TestAssemblyModel modelTest = null;
 
-            Log.WriteLine("----------------------------------");
+            Log.WriteLine();
             Log.WriteLine($"TestEngine: {ricaun.NUnit.TestEngine.Initialize(out string testInitialize)} {testInitialize}");
-            Log.WriteLine("----------------------------------");
+            Log.WriteLine();
 
             foreach (var filePath in Directory.GetFiles(directory, filePattern)) // "*.dll"
             {
@@ -101,51 +100,52 @@ namespace ricaun.RevitTest.Application.Revit
                         TestEngineFilter.CancellationTokenTimeOut = TimeSpan.FromSeconds(3);
 #endif
 
-                        if (FileVersionInfoUtils.TryGetComments(filePath, out ConfigurationComments comments))
+                        var configurationMetadata = ConfigurationMetadata.GetConfigurationMetadata(filePath);
+
+                        if (configurationMetadata.Timeout > 0)
                         {
-                            if (comments.TimeOut > 0)
+                            TestEngineFilter.CancellationTokenTimeOut = TimeSpan.FromMinutes(configurationMetadata.Timeout);
+                            Log.WriteLine($"Tasks.Timeout: {configurationMetadata.Timeout} minutes");
+                        }
+
+                        string containTestNameForNoRevitContext = configurationMetadata.Name;
+                        if (string.IsNullOrEmpty(containTestNameForNoRevitContext) == false)
+                        {
+                            Log.WriteLine($"Tasks.Name: {containTestNameForNoRevitContext}");
+                            if (TestEngineFilter.ExplicitEnabled == false)
                             {
-                                TestEngineFilter.CancellationTokenTimeOut = TimeSpan.FromSeconds(comments.TimeOut);
-                                Log.WriteLine($"TimeOut: {comments.TimeOut}");
+                                TestEngineFilter.ExplicitEnabled = false;
+                                TestEngineFilter.TestNames.AddRange(TestEngine.GetTestFullNames(filePath));
                             }
 
-                            string containTestNameForNoRevitContext = comments.TestAsync;
-                            if (string.IsNullOrEmpty(containTestNameForNoRevitContext) == false)
+                            var originalTestNames = TestEngineFilter.TestNames.ToArray();
+                            var testGroupNoContext = originalTestNames
+                                .GroupBy(str => str.Contains(containTestNameForNoRevitContext))
+                                .ToDictionary(group => group.Key, group => group.ToList());
+
+                            if (testGroupNoContext.TryGetValue(false, out var inContextTestNames))
                             {
-                                if (TestEngineFilter.ExplicitEnabled == false)
-                                {
-                                    TestEngineFilter.ExplicitEnabled = false;
-                                    TestEngineFilter.TestNames.AddRange(TestEngine.GetTestFullNames(filePath));
-                                }
-
-                                var originalTestNames = TestEngineFilter.TestNames.ToArray();
-                                var testGroupNoContext = originalTestNames
-                                    .GroupBy(str => str.Contains(containTestNameForNoRevitContext))
-                                    .ToDictionary(group => group.Key, group => group.ToList());
-
-                                if (testGroupNoContext.TryGetValue(false, out var inContextTestNames))
-                                {
-                                    TestEngineFilter.TestNames.Clear();
-                                    TestEngineFilter.TestNames.AddRange(inContextTestNames);
-                                    modelTest = await revitTask.Run(() => TestEngine.TestAssembly(filePath, parameters));
-                                }
-
-                                if (testGroupNoContext.TryGetValue(true, out var testAsyncNames))
-                                {
-                                    Log.WriteLine($"TestAsync: [{string.Join(",", testAsyncNames)}]");
-                                    TestEngineFilter.TestNames.Clear();
-                                    TestEngineFilter.TestNames.AddRange(testAsyncNames);
-                                    var modelTestAsync = TestEngine.TestAssembly(filePath, parameters);
-                                    if (modelTest is null) modelTest = modelTestAsync;
-                                    else
-                                    {
-                                        modelTest.Tests.AddRange(modelTestAsync.Tests);
-                                    }
-                                }
-
+                                Log.WriteLine($"InContext: [{string.Join(",", inContextTestNames)}]");
                                 TestEngineFilter.TestNames.Clear();
-                                TestEngineFilter.TestNames.AddRange(originalTestNames);
+                                TestEngineFilter.TestNames.AddRange(inContextTestNames);
+                                modelTest = await revitTask.Run(() => TestEngine.TestAssembly(filePath, parameters));
                             }
+
+                            if (testGroupNoContext.TryGetValue(true, out var testAsyncNames))
+                            {
+                                Log.WriteLine($"Tasks: [{string.Join(",", testAsyncNames)}]");
+                                TestEngineFilter.TestNames.Clear();
+                                TestEngineFilter.TestNames.AddRange(testAsyncNames);
+                                var modelTestAsync = await Task.Run(() => TestEngine.TestAssembly(filePath, parameters));
+                                if (modelTest is null) modelTest = modelTestAsync;
+                                else
+                                {
+                                    modelTest.Tests.AddRange(modelTestAsync.Tests);
+                                }
+                            }
+
+                            TestEngineFilter.TestNames.Clear();
+                            TestEngineFilter.TestNames.AddRange(originalTestNames);
                         }
 
                         if (modelTest is null)
@@ -156,6 +156,7 @@ namespace ricaun.RevitTest.Application.Revit
                         var passed = modelTest.Success ? "Passed" : "Failed";
                         if (modelTest.TestCount == 0) { passed = "No Tests"; }
 
+                        Log.WriteLine();
                         Log.WriteLine($"{modelTest}\t {passed}");
 
                         var tests = modelTest.Tests.SelectMany(e => e.Tests);
@@ -188,95 +189,29 @@ namespace ricaun.RevitTest.Application.Revit
                 }
             }
 
-            Log.WriteLine("----------------------------------");
+            Log.WriteLine();
 
             return modelTest;
         }
 
-        //public static TestAssemblyModel Execute(string filePath, params object[] parameters)
-        //{
-        //    var zipFile = CopyFilesUsingZipFolder(filePath);
-
-        //    if (zipFile is null)
-        //        return null;
-
-        //    var extractToTempSucess = ZipExtension.ExtractToTempFolder(zipFile, out string zipDestination);
-
-        //    if (extractToTempSucess == false)
-        //        return null;
-
-        //    TestAssemblyModel tests = TestDirectory(zipDestination, parameters);
-
-        //    CopyFilesBackUsingZip(filePath, zipDestination);
-
-        //    return tests;
-        //}
-
-        //private static TestAssemblyModel TestDirectory(string directory, params object[] parameters)
-        //{
-        //    TestAssemblyModel modelTest = null;
-
-        //    Log.WriteLine("----------------------------------");
-        //    Log.WriteLine($"TestEngine: {ricaun.NUnit.TestEngine.Initialize(out string testInitialize)} {testInitialize}");
-        //    Log.WriteLine("----------------------------------");
-
-        //    foreach (var filePath in Directory.GetFiles(directory, "*.dll"))
-        //    {
-        //        var fileName = Path.GetFileName(filePath);
-        //        try
-        //        {
-        //            if (TestEngine.ContainNUnit(filePath))
-        //            {
-        //                //Log.WriteLine($"Test File: {fileName}");
-        //                //foreach (var testName in TestEngine.GetTestFullNames(filePath))
-        //                //{
-        //                //    Log.WriteLine($"\t{testName}");
-        //                //}
-
-        //                TestEngineFilter.CancellationTokenTimeOut = TimeSpan.FromMinutes(1);
-
-        //                modelTest = TestEngine.TestAssembly(filePath, parameters);
-
-        //                var passed = modelTest.Success ? "Passed" : "Failed";
-        //                if (modelTest.TestCount == 0) { passed = "No Tests"; }
-
-        //                Log.WriteLine($"{modelTest}\t {passed}");
-
-        //                var tests = modelTest.Tests.SelectMany(e => e.Tests);
-
-        //                foreach (var test in tests)
-        //                {
-        //                    Log.WriteLine($"\t {test.Time}\t {test}");
-        //                    //Debug.WriteLine($"Debug:\t {test}\t {test.Console.Trim()}");
-        //                }
-
-        //                if (tests.Any() == false)
-        //                {
-        //                    Log.WriteLine($"Error: {modelTest.Message}");
-        //                    try
-        //                    {
-        //                        var ex = new Exception(modelTest.Message.Split('\n').FirstOrDefault());
-        //                        modelTest = TestEngine.Fail(filePath, ex);
-        //                    }
-        //                    catch { }
-        //                }
-        //            }
-        //        }
-        //        catch (Exception ex)
-        //        {
-        //            Log.WriteLine($"Error: {fileName} {ex}");
-        //        }
-        //    }
-
-        //    Log.WriteLine("----------------------------------");
-
-        //    return modelTest;
-        //}
-
-        private class ConfigurationComments
+        private class ConfigurationMetadata
         {
-            public string TestAsync { get; set; }
-            public double TimeOut { get; set; }
+            public string Name { get; set; }
+            /// <summary>
+            /// Timeout in minutes
+            /// </summary>
+            public double Timeout { get; set; }
+
+            public static ConfigurationMetadata GetConfigurationMetadata(string filePath)
+            {
+                var configurationMetadata = new ConfigurationMetadata();
+                try
+                {
+                    MetadataMapper.Map(configurationMetadata, TestEngine.GetAssemblyMetadataAttributes(filePath), "ricaun.RevitTest.Application.Tasks");
+                }
+                catch { }
+                return configurationMetadata;
+            }
         }
     }
 }