+++ ricaun.RevitTest.Command/Command/IRunTestService.cs
@@ -12,7 +12,7 @@ namespace ricaun.RevitTest.Command
             string forceLanguageToRevit = null,
             bool forceToOpenNewRevit = false,
             bool forceToCloseRevit = false,
-            int timeoutMinutes = 0,
+            double timeoutMinutes = 0,
             params string[] testFilters);
     }
 }