+++ ricaun.RevitTest.TestAdapter/Metadatas/MetadataMapper.cs
@@ -8,7 +8,7 @@ namespace ricaun.RevitTest.TestAdapter.Metadatas
 {
     internal static class MetadataMapper
     {
-        public static T Map<T>(T destination, IEnumerable<AssemblyMetadataAttribute> attributes) where T : class
+        public static T Map<T>(T destination, IEnumerable<AssemblyMetadataAttribute> attributes, string prefix = null) where T : class
         {
             if (attributes is null)
                 return destination;
@@ -17,7 +17,7 @@ namespace ricaun.RevitTest.TestAdapter.Metadatas
                 .GroupBy(attr => attr.Key)
                 .ToDictionary(group => group.Key, group => group.Last().Value);
 
-            return MapperKey.Map(destination, metadataDictionary);
+            return MapperKey.Map(destination, metadataDictionary, prefix);
         }
     }
 }